requirements.txt
# Cthulu - Core Dependencies
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0

# Configuration
PyYAML>=6.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Logging
structlog>=23.1.0
pydantic>=2.5.0

# Additional dependencies for Hektor enhancement and backtesting UI

# Flask server for backtesting UI
Flask==3.0.0
Flask-CORS==6.0.0
Flask-SocketIO==5.3.5
python-socketio==5.14.0

# Async support
eventlet==0.40.3

# Optimization
scipy==1.11.4
scikit-optimize==0.9.0

# Data export formats
pyarrow==14.0.1  # For Parquet support



